cloud_functions/news-classifier/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
torch
torchvision
transformers
db-dtypes
